Nullable check for Authentication.Microsoft models
